namebase plugin within hsd
