+++ lib/wallet/http.js
@@ -11,11 +11,9 @@ const assert = require('bsert');
 const path = require('path');
 const {Server} = require('bweb');
 const Validator = require('bval');
-const {BufferSet} = require('buffer-map');
 const base58 = require('bcrypto/lib/encoding/base58');
 const MTX = require('../primitives/mtx');
 const Outpoint = require('../primitives/outpoint');
-const Output = require('../primitives/output');
 const sha256 = require('bcrypto/lib/sha256');
 const rules = require('../covenants/rules');
 const random = require('bcrypto/lib/random');
@@ -31,7 +29,6 @@ const {Resource} = require('../dns/resource');
 const common = require('./common');
 // eslint-disable-next-line no-unused-vars
 const Wallet = require('./wallet');
-const util = require('../utils/util');
 const { nextTick } = require('process');
 
 /**
@@ -215,8 +212,6 @@ class HTTP extends Server {
       await this.wdb.rescan(height);
     });
 
-    // Deep Rescan
-    // this.post('/deeprescan', async (req, res) => {
     // Deep Clean
     this.post('/deepclean', async (req, res) => {
       if (!req.admin) {
@@ -266,36 +261,6 @@ class HTTP extends Server {
       await this.wdb.backup(path);
 
       res.json(200, { success: true });
-    });
-
-    // Clear cache
-    this.del('/cache/:cacheName', (req, res) => {
-      const cacheName = req.params.cacheName;
-      try {
-        for (const wallet of this.wdb.wallets.values()) {
-          wallet.clearCache(cacheName);
-        }
-      } catch (err) {
-        return res.json(400, {error: err.message});
-      }
-
-      return res.json(200, {message: 'ok'});
-    });
-
-    // Clear specific cache key
-    this.del('/cache/:cacheName/:cacheKey', (req, res) => {
-      const cacheName = req.params.cacheName;
-      const cacheKey = req.params.cacheKey;
-
-      try {
-        for (const wallet of this.wdb.wallets.values()) {
-          wallet.clearCache(cacheName, cacheKey);
-        }
-      } catch (err) {
-        return res.json(400, {error: err.message});
-      }
-
-      return res.json(200, {message: 'ok'});
     });
 
     // List wallets
@@ -1216,27 +1181,12 @@ class HTTP extends Server {
       const passphrase = valid.str('passphrase');
       const broadcast = valid.bool('broadcast', true);
       const sign = valid.bool('sign', true);
-      const idempotencyKey = valid.str('idempotencyKey');
 
       assert(name, 'Name is required.');
       assert(broadcast ? sign : true, 'Must sign when broadcasting.');
 
-      if (idempotencyKey) {
-        assert(
-          broadcast && sign,
-          'Must sign and broadcast if using idempotency cache'
-        );
-      }
-
       const options = TransactionOptions.fromValidator(valid);
-      const {result, fromCache} = await req.wallet.withOpenCache(
-        idempotencyKey,
-        () => {
-          return req.wallet.createOpen(name, force, options);
-        }
-      );
-      const mtx = result;
-      res.setHeader('From-Cache', fromCache);
+      const mtx = await req.wallet.createOpen(name, force, options);
 
       if (broadcast) {
         const tx = await req.wallet.sendMTX(mtx, passphrase);
@@ -1247,40 +1197,6 @@ class HTTP extends Server {
         await req.wallet.sign(mtx, passphrase);
 
       return res.json(200, mtx.getJSON(this.network));
-    });
-
-    // Batch Open
-    this.post('/wallet/:id/batch/open', async (req, res) => {
-      const valid = Validator.fromRequest(req);
-      const names = valid.array('names');
-      const force = valid.bool('force', false);
-      const passphrase = valid.str('passphrase');
-      const broadcast = valid.bool('broadcast', true);
-      const sign = valid.bool('sign', true);
-      const MAX_NAME_ARRAY_LENGTH = 200;
-
-      assert(names && names.length > 0, 'Names are required.');
-      assert(names.length <= MAX_NAME_ARRAY_LENGTH,
-        `Names array shoud not exceed ${MAX_NAME_ARRAY_LENGTH}`);
-      assert(broadcast ? sign : true, 'Must sign when broadcasting.');
-
-      const options = TransactionOptions.fromValidator(valid);
-      const {mtx, errors, isAllError} = await req.wallet
-                                        .createBatchOpen(names, force, options);
-
-      if (isAllError) // no valid output in mtx
-        return res.json(500, {errors: errors});
-
-      if (broadcast) {
-        const tx = await req.wallet.sendMTX(mtx, passphrase);
-        return res.json(200, {tx: tx.getJSON(this.network), errors: errors});
-      }
-
-      if (sign) {
-        await req.wallet.sign(mtx, passphrase);
-      }
-
-      return res.json(200, {tx: mtx.getJSON(this.network), errors: errors});
     });
 
     // Create Bid
@@ -1292,29 +1208,14 @@ class HTTP extends Server {
       const passphrase = valid.str('passphrase');
       const broadcast = valid.bool('broadcast', true);
       const sign = valid.bool('sign', true);
-      const idempotencyKey = valid.str('idempotencyKey');
 
       assert(name, 'Name is required.');
       assert(bid != null, 'Bid is required.');
       assert(lockup != null, 'Lockup is required.');
       assert(broadcast ? sign : true, 'Must sign when broadcasting.');
-      if (idempotencyKey) {
-        assert(
-          broadcast && sign,
-          'Must sign and broadcast if using idempotency cache'
-        );
-      }
 
       const options = TransactionOptions.fromValidator(valid);
-
-      const {result, fromCache} = await req.wallet.withBidCache(
-        idempotencyKey,
-        () => {
-          return req.wallet.createBid(name, bid, lockup, options);
-        }
-      );
-      const mtx = result;
-      res.setHeader('From-Cache', fromCache);
+      const mtx = await req.wallet.createBid(name, bid, lockup, options);
 
       if (broadcast) {
         const tx = await req.wallet.sendMTX(mtx, passphrase);
@@ -1325,88 +1226,6 @@ class HTTP extends Server {
         await req.wallet.sign(mtx, passphrase);
 
       return res.json(200, mtx.getJSON(this.network));
-    });
-
-    // Batch Bid
-    this.post('/wallet/:id/batch/bid', async (req, res) => {
-      const MAX_BIDS_ARRAY_LENGTH = 200;
-      const valid = Validator.fromRequest(req);
-      const passphrase = valid.str('passphrase');
-
-      if (!valid.has('bids'))
-        throw new Error('Mandatory Bids parameter is missing!');
-
-      let bids = valid.array('bids');
-      assert(bids.length > 0 && bids.length <= MAX_BIDS_ARRAY_LENGTH,
-        `Bids are required to be non empty and shoud not exceed
-        ${MAX_BIDS_ARRAY_LENGTH}`);
-
-      bids.forEach((element) => {
-        const validator = new Validator(element);
-        if (!validator.has('name') ||
-            !validator.has('bid')  ||
-            !validator.has('lockup') ||
-            !validator.has('idempotencyKey')) {
-          throw new Error('Bids contains a Bid with missing mandatory fields!');
-        } else { // check types
-          validator.str('name');
-          validator.u64('bid');
-          validator.u64('lockup');
-          validator.str('idempotencyKey');
-        }
-      });
-
-     // add value property (same as bid) to each bid
-     // value property is same with bid property and used internally
-      bids = bids.map((element) => {
-        element['value'] = element.bid;
-        return element;
-      });
-
-      const options = TransactionOptions.fromValidator(valid);
-
-      const uniqueBids = [];
-      const bidResults = [];
-
-      let errors  = [];
-
-      const bidCache = req.wallet.sendBidResults;
-
-      for (const bid of bids) {
-        const bidResultFromCache = bidCache.get(bid.idempotencyKey);
-        if (bidResultFromCache)
-          bidResults.push(bidResultFromCache);
-        else
-          uniqueBids.push(bid);
-      }
-
-      if (uniqueBids.length > 0) {
-        const {mtx, errorMessages} = await req.wallet
-          .createBatchBid(uniqueBids, options);
-
-        errors = errorMessages;
-
-        // always broadcast and sign
-        const tx = await req.wallet.sendMTX(mtx, passphrase);
-        const txJSON = tx.getJSON(this.network);
-
-        const processedBids = util
-          .postProcessBatchBids(txJSON.hash, txJSON.outputs, mtx.outputs);
-
-        // update cache
-        processedBids.forEach((bid) => {
-          const cachedBid = Object.assign({}, bid);
-          cachedBid.fromCache = true;
-          bidCache.set(bid.idempotency_key, cachedBid);
-        });
-
-        processedBids.forEach(bid => bidResults.push(bid));
-      }
-
-      return res.json(200, {
-        processedBids: bidResults,
-        errorMessages: errors
-      });
     });
 
     // Create auction-related transactions in advance (bid and reveal for now)
@@ -1477,158 +1296,6 @@ class HTTP extends Server {
       return res.json(200, mtx.getJSON(this.network));
     });
 
-    // Create Batch Reveal
-    this.post('/wallet/:id/batch/reveal', async (req, res) => {
-      const valid = Validator.fromRequest(req);
-      const names = valid.array('names');
-      const passphrase = valid.str('passphrase');
-      const broadcast = valid.bool('broadcast', true);
-      const sign = valid.bool('sign', true);
-
-      assert(names && names.length > 0, 'Names are required.');
-      assert(broadcast ? sign : true, 'Must sign when broadcasting.');
-
-      const options = TransactionOptions.fromValidator(valid);
-      const {mtx, errorMessages} = await req.wallet
-        .createBatchReveal(names, options);
-
-      if (broadcast) {
-        const tx = await req.wallet.sendMTX(mtx, passphrase);
-        return res.json(200, {tx: tx.getJSON(this.network),
-          errors: errorMessages});
-      }
-
-      if (sign)
-        await req.wallet.sign(mtx, passphrase);
-
-      return res.json(200, {tx: mtx.getJSON(this.network),
-        errorMessages: errorMessages});
-    });
-
-    // Create Batch Reveal
-    this.post('/wallet/:id/batch/revealwithcache', async (req, res) => {
-      const valid = Validator.fromRequest(req);
-      const names = valid.array('names');
-      const passphrase = valid.str('passphrase');
-
-      assert(names && names.length > 0, 'Names are required.');
-      const options = TransactionOptions.fromValidator(valid);
-
-      const revealCache = req.wallet.sendRevealResults;
-
-      const { cacheMisses: uniqueNames, cacheHits: processedReveals } = util
-        .retrieveMultipleFromCache(revealCache, names);
-
-      if (uniqueNames.length === 0) {
-        return res.json(200, {processedReveals, errors: []});
-      }
-
-      const batchRevealResponse = await req.wallet
-        .createBatchReveal(uniqueNames, options);
-
-      const mtx = batchRevealResponse.mtx;
-      const errorMessages = batchRevealResponse.errorMessages;
-
-      // always broadcast
-      const tx = await req.wallet.sendMTX(mtx, passphrase);
-      const txJSON = tx.getJSON(this.network);
-
-      const nameHashMap = new Map();
-      for (const name of uniqueNames) {
-        // to prevent a weird case of numerics (0)
-        const nameHash = rules.hashName(String(name)).toString('hex');
-        nameHashMap.set(nameHash, name);
-      }
-
-      const {outputs, hash: txHash} = txJSON;
-
-      outputs.forEach((output, index) => {
-        if (output.covenant.type === rules.types.REVEAL) {
-          const nameHash = output.covenant.items[0];
-          const name = nameHashMap.get(nameHash);
-          const processedReveal = util
-            .postProcessOutput(output, index, txHash, name);
-          processedReveals.push(processedReveal);
-          util.storeMultipleInCache(processedReveal, revealCache);
-        }
-      });
-
-      return res.json(200, {processedReveals, errors: errorMessages});
-    });
-
-    // Create Batch Finish (Redeem or Register)
-    this.post('/wallet/:id/batch/finish', async (req, res) => {
-      const valid = Validator.fromRequest(req);
-      const finishRequests = valid.array('finishRequests');
-      const passphrase = valid.str('passphrase');
-
-      assert(finishRequests && finishRequests.length > 0,
-        'FinishRequests are required.');
-
-      finishRequests.forEach((element) => {
-        const validator = new Validator(element);
-        if (!validator.has('name') || !validator.has('data'))
-          throw new Error('name and data must be present in every element.');
-
-        validator.str('name');
-        validator.obj('data');
-
-        const dataValidator = new Validator(element.data);
-        dataValidator.array('records');
-      });
-
-      const formattedFinishRequests = finishRequests.map(({name, data}) => {
-        return {name, data: Resource.fromJSON(data)};
-      });
-
-      const options = TransactionOptions.fromValidator(valid);
-      const finishCache = req.wallet.sendFinishResults;
-
-      const uniqueNames = [];
-      const processedFinishes = [];
-
-      formattedFinishRequests.forEach(({name, data}) => {
-        const cachedResponse = finishCache.get(name);
-        if (!cachedResponse) {
-          uniqueNames.push({name, data});
-        } else {
-          cachedResponse.forEach(cachedName =>
-            processedFinishes.push(cachedName));
-        }
-      });
-
-      let errorMessages = [];
-      if (uniqueNames.length > 0) {
-        const nameHashMap = new Map();
-        uniqueNames.forEach(({name, data}) => {
-          const nameHash = rules.hashName(name).toString('hex');
-          nameHashMap.set(nameHash, name);
-        });
-
-        // always broadcast
-        const finishResponse = await req.wallet.createBatchFinish(
-          uniqueNames,
-          options
-        );
-        errorMessages = finishResponse.errorMessages;
-
-        const tx = await req.wallet.sendMTX(finishResponse.mtx, passphrase);
-        const txJSON = tx.getJSON(this.network);
-
-        util.postProcessBatchFinishes(
-          txJSON,
-          finishCache,
-          nameHashMap,
-          processedFinishes
-        );
-      }
-
-      return res.json(200, {
-        processedFinishes,
-        errorMessages
-      });
-    });
-
     // Create Redeem
     this.post('/wallet/:id/redeem', async (req, res) => {
       const valid = Validator.fromRequest(req);
@@ -1687,7 +1354,7 @@ class HTTP extends Server {
 
       const options = TransactionOptions.fromValidator(valid);
 
-      const {result, fromCache} = await req.wallet.withBidCache(
+      const {result, fromCache} = await req.wallet.withUpdateCache(
         idempotencyKey,
         () => {
           return req.wallet.createUpdate(name, resource, options);
@@ -1878,103 +1545,6 @@ class HTTP extends Server {
       return res.json(200, mtx.getJSON(this.network));
     });
 
-    // SendMany
-    this.post('/wallet/:id/sendmany', async (req, res) => {
-      const valid = Validator.fromRequest(req);
-      const account = valid.str('account', 'default');
-      const sendTo = valid.array('sendto', []);
-      const minconf = valid.u32('minconf', 1);
-      const subtractFee = valid.bool('subtractfee', false);
-      const passphrase = valid.str('passphrase');
-
-      if (sendTo.length === 0)
-        throw new Error('parameter sendto is required and can not be empty!');
-
-      const cache = req.wallet.hsdTransferResults;
-      const results = [];
-
-      for (const {idempotency_key: idempotencyKey} of sendTo) {
-        if (idempotencyKey && cache.has(idempotencyKey))
-          results.push(cache.get(idempotencyKey));
-      }
-
-      const filteredSendTo = sendTo.filter(element =>
-        !cache.has(element.idempotency_key));
-
-      const addressToIdempotencyKeyMap = new Map();
-      const uniq = new BufferSet();
-
-      const outputs = filteredSendTo.map((element) => {
-        const to = new Validator(element);
-
-        const idempotencyKey = to.str('idempotency_key');
-        if (!idempotencyKey)
-          throw new Error('idempotencyKey is missing!');
-
-        const value   = to.u64('value');
-        if (!value)
-          throw new Error(`value: ${value} is invalid!`);
-
-        const rawAddress = to.str('address');
-        const addr = parseAddress(rawAddress, this.network);
-        addressToIdempotencyKeyMap.set(addr.toString(), {
-          idempotencyKey: idempotencyKey,
-          rawAddress: rawAddress
-        });
-
-        const hash = addr.getHash();
-
-        if (uniq.has(hash))
-          throw new Error('Invalid parameter!');
-
-          uniq.add(hash);
-
-          const output = new Output();
-          output.value = value;
-          output.address = addr;
-
-          return output;
-      });
-
-      if (outputs.length > 0) {
-        const options = {
-          outputs: outputs,
-          subtractFee: subtractFee,
-          account: account,
-          depth: minconf
-        };
-
-        const tx = await req.wallet.send(options, passphrase);
-
-        const txJSON = tx.toJSON();
-        const txOutputsLen = txJSON.outputs.length;
-
-        for (let i=0; i<txOutputsLen; i++) {
-          const output = txJSON.outputs[i];
-          if (addressToIdempotencyKeyMap.has(output.address)) {
-            const {idempotencyKey} = addressToIdempotencyKeyMap
-              .get(output.address);
-
-            const hsdTransfer =
-            {
-              idempotency_key: idempotencyKey,
-              tx_hash: txJSON.hash,
-              output_index: i,
-              output: output
-            };
-
-            results.push(hsdTransfer);
-
-            const cachedHsdTransfer = Object.assign({}, hsdTransfer);
-            cachedHsdTransfer.fromCache = true;
-            cache.set(cachedHsdTransfer.idempotency_key, cachedHsdTransfer);
-          }
-        }
-      }
-
-      return res.json(200, {processedWithdrawals: results});
-    });
-
     // resync wallet db
     this.post('/wallet/:id/resync', async (req, res) => {
       const valid = Validator.fromRequest(req);
@@ -2155,6 +1725,38 @@ class HTTP extends Server {
       socket.leave(`w:${id}`);
 
       return null;
+    });
+  }
+
+  /**
+   * Registers a new http handler at provided path
+   * the handler method is passed regular req, res along
+   * with a context object that holds vallet specific
+   * objects that might be used to extend wallet plugin
+   * functionality
+   *
+   * @param {*} httpMethod string value of one of
+   * 'post', 'get', 'put', 'patch', 'delete'
+   * @param {*} path location handler is listening
+   * @param {*} handler http handler that will handle requests
+   * arriving to `path` registered
+   * @param {Object} context context
+   * @returns {undefined}
+   */
+   registerRoute(httpMethod, path, handler, context) {
+    const validHttpMethods = ['post', 'get', 'put', 'patch', 'del'];
+    assert(httpMethod && typeof httpMethod === 'string',
+    'httpMethod parameter must be provided and it must be a string');
+    assert(validHttpMethods.includes(httpMethod.toLowerCase()),
+    'httpMethod must be one of ' + validHttpMethods.join(','));
+
+    // update context
+    context.TransactionOptions = TransactionOptions;
+    context.Validator = Validator;
+    context.network = this.network;
+
+    this[httpMethod](path, async (req, res) => {
+      await handler(req, res, context);
     });
   }
 }
