+++ lib/wallet/rpc.js
@@ -202,9 +202,6 @@ class RPC extends RPCBase {
     this.add('createcancel', this.createCancel);
     this.add('createfinalize', this.createFinalize);
     this.add('createrevoke', this.createRevoke);
-    this.add('sendbididempotent', this.sendBidIdempotent);
-    this.add('sendopenidempotent', this.sendOpenIdempotent);
-    this.add('sendupdateidempotent', this.sendUpdateIdempotent);
 
     // Compat
     this.add('getauctions', this.getNames);
@@ -2030,41 +2027,6 @@ class RPC extends RPCBase {
     return {name, force, account};
   }
 
-  async sendOpenIdempotent(args, help) {
-    const opts = this._validateOpenIdempotent(args, help, 'sendopenidempotent');
-    const wallet = this.wallet;
-    const openResult = await wallet.sendOpenIdempotent(
-      opts.name,
-      opts.force,
-      opts.idempotencyKey,
-      { account: opts.account }
-    );
-    const tx = openResult.result;
-    const fromCache = openResult.fromCache;
-
-    const result = tx.getJSON(this.network);
-    result.fromCache = fromCache;
-    return result;
-  }
-
-  _validateOpenIdempotent(args, help, method) {
-    const msg = `${method} "name" idempotencykey ( force "account" )`;
-
-    if (help || args.length < 2 || args.length > 4)
-      throw new RPCError(errs.MISC_ERROR, msg);
-
-    const valid = new Validator(args);
-    const name = valid.str(0);
-    const idempotencyKey = valid.str(1);
-    const force = valid.bool(2, false);
-    const account = valid.str(3);
-
-    if (!name || !rules.verifyName(name))
-      throw new RPCError(errs.TYPE_ERROR, 'Invalid name.');
-
-    return { name, force, idempotencyKey, account };
-  }
-
   async sendBid(args, help) {
     const opts = this._validateBid(args, help, 'sendbid');
     const wallet = this.wallet;
@@ -2111,48 +2073,6 @@ class RPC extends RPCBase {
       name,
       bid,
       value,
-      account
-    };
-  }
-
-  async sendBidIdempotent(args, help) {
-    const opts = this._validateBidIdempotent(args, help, 'sendbididempotent');
-    const wallet = this.wallet;
-    const tx = await wallet.sendBid(opts.name, opts.bid, opts.value, {
-      account: opts.account,
-      idempotencyKey: opts.idempotencyKey
-    });
-
-    return tx.getJSON(this.network);
-  }
-
-  _validateBidIdempotent(args, help, method) {
-    const msg = `${method} "name" bid value idempotencykey ( "account" )`;
-
-    if (help || args.length < 4 || args.length > 5)
-      throw new RPCError(errs.MISC_ERROR, msg);
-
-    const valid = new Validator(args);
-    const name = valid.str(0);
-    const bid = valid.ufixed(1, EXP);
-    const value = valid.ufixed(2, EXP);
-    const idempotencyKey = valid.str(3);
-    const account = valid.str(4);
-
-    if (!name || !rules.verifyName(name))
-      throw new RPCError(errs.TYPE_ERROR, 'Invalid name.');
-
-    if (bid == null || value == null)
-      throw new RPCError(errs.TYPE_ERROR, 'Invalid values.');
-
-    if (bid > value)
-      throw new RPCError(errs.TYPE_ERROR, 'Invalid bid.');
-
-    return {
-      name,
-      bid,
-      value,
-      idempotencyKey,
       account
     };
   }
@@ -2267,22 +2187,6 @@ class RPC extends RPCBase {
     const wallet = this.wallet;
     const tx = await wallet.sendUpdate(opts.name, opts.resource, {
       account: opts.account
-    });
-
-    return tx.getJSON(this.network);
-  }
-
-  async sendUpdateIdempotent(args, help) {
-    const opts = this._validateUpdateIdempotent(
-      args,
-      help,
-      'sendupdateidempotent'
-    );
-
-    const wallet = this.wallet;
-    const tx = await wallet.sendUpdate(opts.name, opts.resource, {
-      account: opts.account,
-      idempotencyKey: opts.idempotencyKey
     });
 
     return tx.getJSON(this.network);
