+++ lib/wallet/wallet.js
@@ -1849,25 +1849,6 @@ class Wallet extends EventEmitter {
     }
   }
 
-  /**
-   * Create and send an open with an idempotency key
-   * MTX with a lock.
-   * @param {String} name
-   * @param {Boolean} force
-   * @param {Object} options
-   */
-
-  async sendOpenIdempotent(name, force, idempotencyKey, options) {
-    const unlock = await this.fundLock.lock();
-    try {
-      return await this.withOpenCache(idempotencyKey,
-        () => this._sendOpen(name, force, options)
-      );
-    } finally {
-      unlock();
-    }
-  }
-
    /**
    * Make a bid MTX.
    * @param {String} name
@@ -2073,22 +2054,6 @@ class Wallet extends EventEmitter {
     return {result: result, fromCache: false};
   }
 
-  async withBidCache(idempotencyKey, bidCallback) {
-    return await Wallet.doWithCache(
-      this.sendBidResults,
-      idempotencyKey,
-      bidCallback
-    );
-  }
-
-  async withOpenCache(idempotencyKey, openCallback) {
-    return await Wallet.doWithCache(
-      this.sendOpenResults,
-      idempotencyKey,
-      openCallback
-    );
-  }
-
   async withUpdateCache(idempotencyKey, updateCallback) {
     return await Wallet.doWithCache(
       this.sendUpdateResults,
@@ -2115,20 +2080,16 @@ class Wallet extends EventEmitter {
 
   /**
    * Create and send a bid MTX.
-   * @param {String} name the name to bid on
-   * @param {Number} value the value of the bid
-   * @param {Number} lockup the total lockup, including blind
-   * @param {Object} options bid options
-   * @returns {Object} a bid result
+   * @param {String} name
+   * @param {Number} value
+   * @param {Number} lockup
+   * @param {Object} options
    */
+
   async sendBid(name, value, lockup, options) {
     const unlock = await this.fundLock.lock();
     try {
-      const idempotencyKey = options ? options.idempotencyKey : null;
-      const {result} = await this.withBidCache(idempotencyKey,
-        () => this._sendBid(name, value, lockup, options)
-      );
-      return result;
+      return await this._sendBid(name, value, lockup, options);
     } finally {
       unlock();
     }
